+++ base/src/ring/mod.rs
@@ -118,7 +118,17 @@ pub trait ExtendedGcd<Rhs = Self> {
     fn gcd_ext(self, rhs: Rhs) -> (Self::OutputGcd, Self::OutputCoeff, Self::OutputCoeff);
 }
 
-/// Compute the roots (square root, cubic root and arbitrary n-th roots)
+
+// TODO: more docs
+/// Compute the roots (square root, cubic root) of an integer.
+pub trait Root {
+    type Output;
+
+    fn sqrt(&self) -> Self::Output;
+    fn cbrt(&self) -> Self::Output;
+}
+
+/// Compute the roots (square root, cubic root) of an integer, and returns the remainder of the root as well.
 pub trait RootRem {
     type Output; // TODO(v0.3): Separate root type and remainder type
 
@@ -126,6 +136,7 @@ pub trait RootRem {
     #[deprecated(note = "this function is never supported and will be removed in the next version")]
     fn nth_root_rem(self, n: usize) -> (Self::Output, Self::Output);
 
+    // TODO(v0.3): taking reference instead
     fn sqrt_rem(self) -> (Self::Output, Self::Output);
     fn cbrt_rem(self) -> (Self::Output, Self::Output);
 }