+++ integer/src/lib.rs
@@ -101,6 +101,7 @@ mod gcd;
 mod gcd_ops;
 mod helper_macros;
 mod ibig;
+mod iter;
 mod log;
 mod math;
 mod memory;
@@ -114,17 +115,14 @@ mod primitive;
 mod radix;
 mod remove;
 mod repr;
+mod root;
+mod root_ops;
 mod shift;
 mod shift_ops;
 mod sign;
 mod sqr;
+mod third_party;
 mod ubig;
 
-#[cfg(feature = "rand")]
-pub mod rand;
-
-#[cfg(feature = "num-traits")]
-mod num_traits;
-
-#[cfg(feature = "serde")]
-mod serde;
+// All the public items from third_party will be exposed
+pub use third_party::*;